+++ BUILD.gn
@@ -1190,7 +1190,6 @@ config("grpc_config") {
         "include/grpcpp/support/async_unary_call_impl.h",
         "include/grpcpp/support/byte_buffer.h",
         "include/grpcpp/support/channel_arguments.h",
-        "include/grpcpp/support/channel_arguments_impl.h",
         "include/grpcpp/support/client_callback.h",
         "include/grpcpp/support/client_callback_impl.h",
         "include/grpcpp/support/client_interceptor.h",
