+++ CMakeLists.txt
@@ -2833,7 +2833,6 @@ foreach(_hdr
   include/grpcpp/support/async_unary_call_impl.h
   include/grpcpp/support/byte_buffer.h
   include/grpcpp/support/channel_arguments.h
-  include/grpcpp/support/channel_arguments_impl.h
   include/grpcpp/support/client_callback.h
   include/grpcpp/support/client_callback_impl.h
   include/grpcpp/support/client_interceptor.h
@@ -3525,7 +3524,6 @@ foreach(_hdr
   include/grpcpp/support/async_unary_call_impl.h
   include/grpcpp/support/byte_buffer.h
   include/grpcpp/support/channel_arguments.h
-  include/grpcpp/support/channel_arguments_impl.h
   include/grpcpp/support/client_callback.h
   include/grpcpp/support/client_callback_impl.h
   include/grpcpp/support/client_interceptor.h
