+++ Makefile
@@ -4957,7 +4957,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/support/async_unary_call_impl.h \
     include/grpcpp/support/byte_buffer.h \
     include/grpcpp/support/channel_arguments.h \
-    include/grpcpp/support/channel_arguments_impl.h \
     include/grpcpp/support/client_callback.h \
     include/grpcpp/support/client_callback_impl.h \
     include/grpcpp/support/client_interceptor.h \
@@ -5647,7 +5646,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/support/async_unary_call_impl.h \
     include/grpcpp/support/byte_buffer.h \
     include/grpcpp/support/channel_arguments.h \
-    include/grpcpp/support/channel_arguments_impl.h \
     include/grpcpp/support/client_callback.h \
     include/grpcpp/support/client_callback_impl.h \
     include/grpcpp/support/client_interceptor.h \
