+++ build_autogenerated.yaml
@@ -2165,7 +2165,6 @@ libs:
   - include/grpcpp/support/async_unary_call_impl.h
   - include/grpcpp/support/byte_buffer.h
   - include/grpcpp/support/channel_arguments.h
-  - include/grpcpp/support/channel_arguments_impl.h
   - include/grpcpp/support/client_callback.h
   - include/grpcpp/support/client_callback_impl.h
   - include/grpcpp/support/client_interceptor.h
@@ -2549,7 +2548,6 @@ libs:
   - include/grpcpp/support/async_unary_call_impl.h
   - include/grpcpp/support/byte_buffer.h
   - include/grpcpp/support/channel_arguments.h
-  - include/grpcpp/support/channel_arguments_impl.h
   - include/grpcpp/support/client_callback.h
   - include/grpcpp/support/client_callback_impl.h
   - include/grpcpp/support/client_interceptor.h
