+++ gRPC-C++.podspec
@@ -177,7 +177,6 @@ Pod::Spec.new do |s|
                       'include/grpcpp/support/async_unary_call_impl.h',
                       'include/grpcpp/support/byte_buffer.h',
                       'include/grpcpp/support/channel_arguments.h',
-                      'include/grpcpp/support/channel_arguments_impl.h',
                       'include/grpcpp/support/client_callback.h',
                       'include/grpcpp/support/client_callback_impl.h',
                       'include/grpcpp/support/client_interceptor.h',
