+++ tools/doxygen/Doxyfile.c++
@@ -1037,7 +1037,6 @@ include/grpcpp/support/async_unary_call.h \
 include/grpcpp/support/async_unary_call_impl.h \
 include/grpcpp/support/byte_buffer.h \
 include/grpcpp/support/channel_arguments.h \
-include/grpcpp/support/channel_arguments_impl.h \
 include/grpcpp/support/client_callback.h \
 include/grpcpp/support/client_callback_impl.h \
 include/grpcpp/support/client_interceptor.h \
